adjust wait script
